- Separate join test in file test_join.py - Remove default value of _join() parameters - Change targets list name for better identification - Make cond, how and on parameters of _join() not Optional
